+++ intelmq/lib/upgrades.py
@@ -338,6 +338,7 @@ def v213_feed_changes(defaults, runtime, harmonization, dry_run):
     found_zeus = []
     found_bitcash = []
     found_ddos_attack = []
+    found_ransomware = []
     changed = None
     messages = []
     for bot_id, bot in runtime.items():
@@ -349,18 +350,23 @@ def v213_feed_changes(defaults, runtime, harmonization, dry_run):
                 found_zeus.append(bot_id)
             elif bot["parameters"].get("http_url").startswith("https://bitcash.cz/misc/log/blacklist"):
                 found_bitcash.append(bot_id)
+            elif bot["parameters"].get("http_url").startswith("https://ransomwaretracker.abuse.ch/feeds/csv/"):
+                found_ransomware.append(bot_id)
         if bot["module"] == "intelmq.bots.collectors.http.collector_http_stream":
             if bot["parameters"].get("http_url").startswith("https://feed.caad.fkie.fraunhofer.de/ddosattackfeed"):
                 found_ddos_attack.append(bot_id)
     if found_zeus:
         messages.append('A discontinued feed "Zeus Tracker" has been found '
-                        'as bot %s. Remove it yourself please.' % ', '.join(found_zeus))
+                        'as bot %s. Remove it yourself please.' % ', '.join(sorted(found_zeus)))
     if found_bitcash:
         messages.append('The discontinued feed "Bitcash.cz" has been found '
-                        'as bot %s. Remove it yourself please.' % ', '.join(found_bitcash))
+                        'as bot %s. Remove it yourself please.' % ', '.join(sorted(found_bitcash)))
     if found_ddos_attack:
         messages.append('The discontinued feed "Fraunhofer DDos Attack" has been found '
-                        'as bot %s. Remove it yourself please.' % ', '.join(found_ddos_attack))
+                        'as bot %s. Remove it yourself please.' % ', '.join(sorted(found_ddos_attack)))
+    if found_ransomware:
+        messages.append('The discontinued feed "Abuse.ch Ransomware Tracker" has been found '
+                        'as bot %s. Remove it yourself please.' % ', '.join(sorted(found_ransomware)))
     messages = ' '.join(messages)
     return messages if messages else changed, defaults, runtime, harmonization